etl_scripts/requirements.txt
pandas
numpy
openpyxl
python-dotenv
db-dtypes
jupyter
notebook
ipykernel
seaborn
ipywidgets
scikit-learn
xgboost
google-cloud-bigquery
